+++ packages/queries/src/stackscripts/index.ts
@@ -1,2 +1,3 @@
 export * from './requests';
 export * from './stackscripts';
+export * from './keys';
